+++ NEWS.md
@@ -11,6 +11,7 @@
   in this connection `as_r()` was renamed to `as_expr()` and 
   `as_symbol()` to `as_sym()`. Also, `as_sym()` changed argument from `declare_variables` to `declare_symbols`.
 * Changed internals such that `der()`, `der2()` and `solve_sys()` now takes multiple variables with `list()` instead of `c()`; see also `matrify()`
+* Added `diag_sym()` and `matrix_sym()`
 * Call SymPy functions directy with `sympy_func()`
 * Minor bugs fixed